requirements_prod.txt
colorama
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.1
tensorflow==2.10.0
google-cloud-bigquery
google-cloud-storage
db-dtypes
tqdm
regex

mlflow==2.1.1

fastapi
pytz
uvicorn
uuid
python-multipart

python-dotenv
